test: Moves helpers.py to top-level test directory
